Consider function depth when resolving instance argument name
To account for nested function scoping, determine the depth of the current node and use the one whose parent context is the containing `ClassDef` node. This should hopefully end up being the function that accepts the instance variable.

This isn't perfect, e.g. things might go badly if the nested function accepts the instance under a new name, but I think it should do well for most projects?

See: #3